requirements.txt
asn1crypto==0.24.0
certifi==2018.1.18
cffi==1.11.4
click==6.7
cryptography==2.1.4
Flask==0.12.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
prometheus-client==0.1.1
pycparser==2.18
pyOpenSSL==17.5.0
six==1.11.0
Werkzeug==0.14.1
